redis进行问卷以及回答详情等内容的缓存。 AOP和@CleanCache注解，解决缓存一致性。 编写AOP切面类，打印请求的日志。 使用布隆过滤器进行有效问卷的缓存，防止缓存穿透。

+++ src/main/java/SurveySystem/service/Impl/PermissionServiceImpl.java
@@ -1,8 +1,10 @@
 package SurveySystem.service.Impl;
 
 import SurveySystem.entity.Permission;
+import SurveySystem.entity.Role;
 import SurveySystem.mapper.PermissionMapper;
 import SurveySystem.service.PermissionService;
+import SurveySystem.service.RolePermissionService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -13,6 +15,8 @@ public class PermissionServiceImpl implements PermissionService {
 
     @Autowired
     private PermissionMapper permissionMapper;
+    @Autowired
+    private RolePermissionService rolePermissionService;
 
     @Override
     public int insertPermission(Permission permission) {
@@ -20,8 +24,14 @@ public class PermissionServiceImpl implements PermissionService {
     }
 
     @Override
-    public void deletePermissionById(int id) {
+    public Boolean deletePermissionById(int id) {
+        //检查是否有其他角色在用此权限。
+        List<String> roleIds = rolePermissionService.getRoleIdsByPermissionId(id);
+        if(roleIds!=null&&roleIds.size()>0){
+            return false;
+        }
         permissionMapper.deletePermissionById(id);
+        return true;
     }
 
     @Override
